+++ bin/install.mjs
@@ -20,7 +20,7 @@ if (process.argv.length < 3) {
 const projectName = process.argv[2]
 const currentPath = process.cwd()
 const projectPath = path.join(currentPath, projectName)
-const git_repo = 'https://github.com/praghus/platfuse-example.git'
+const git_repo = 'https://github.com/praghus/platfuse-boilerplate.git'
 
 // create project directory
 if (fs.existsSync(projectPath)) {