+++ CHANGELOG.md
@@ -4,6 +4,34 @@ All notable changes to the Platfuse Game Engine project will be documented in th
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.2] - unreleased
+
+### Added
+
+-   N/A
+
+### Changed
+
+-   Text printing method (optional stroke).
+-   Entity overlaping metod allow to check colissions also with points (vec2).
+
+### Deprecated
+
+-   N/A
+
+### Removed
+
+-   N/A
+
+### Fixed
+
+-   Camera scroll alnd scale.
+-   Rotations for non animated objects.
+
+### Security
+
+-   N/A
+
 ## [1.3.0] - 2024-07-04
 
 ### Added